render TeamPage when updated
