+++ src/pages/Team/Overview/hooks.ts
@@ -1,0 +1,7 @@
+import { useRecoilValue } from 'recoil';
+import { teamState } from '../hooks';
+
+export const useTeam = (teamId: string) => {
+  const team = useRecoilValue(teamState(teamId));
+  return { team };
+};
